# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/v1/classes": {
            "get": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Get list of classes",
                "description": "Returns list of classes",
                "operationId": "getClassesList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Store new class",
                "description": "Stores a new class and returns the created class",
                "operationId": "storeClass",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "school_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "JSS1"
                                    },
                                    "school_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9a7a7b0e-0b1c-4f0e-8c1a-0b1c4f0e8c1a"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/classes/{id}": {
            "get": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Get session information",
                "description": "Returns class data",
                "operationId": "getClassById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Update existing class",
                "description": "Updates a class and returns the updated class",
                "operationId": "updateClass",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "JSS1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Delete existing class",
                "description": "Deletes a class and returns no content",
                "operationId": "deleteClass",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/classes/{classId}/arms": {
            "get": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Get list of class arms",
                "description": "Returns list of class arms for a given class",
                "operationId": "getClassArmsList",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "Class id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.2"
                ],
                "summary": "Store new class arm",
                "description": "Stores a new class arm and returns the created class arm",
                "operationId": "storeClassArm",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "Class id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}